requirements.txt
Django==2.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2018.3
six==1.11.0
